==> 1_code/7_extracting_transect_depth_profiles.Rmd <==
---
title: "7. Extracting Transect Depth Profiles"
output:
  html_document:
    toc: true
    number_sections: true
    toc_float: true
    theme: cerulean
    df_print: paged
    code_folding: "hide"
---

```{r Load Data}
library(sf)
library(raster)
library(here)
library(tidyverse)

Hawaii_Bathymetry <- raster(here("2_output", "modified_rasters", "Hawaii_Bathymetry_-30_to_1.tif")) ##Choose which file you want to use for depths. I will use the -30 to 1 m file and then later remove offshore depths greater than 7 m deep.
  
  
Hawaii_Transects <- st_read(here("2_output", "modified_shapefiles","joined_databases" , "reefs_and_erosion.gpkg"))
Hawaii_Transects<-Hawaii_Transects%>%
  dplyr::select(Transects_ID,island_reef) %>%
  distinct()

st_write(Hawaii_Transects,here("test_distict.gpkg"), delete_layer=TRUE)

{plot(Hawaii_Bathymetry)
plot(Hawaii_Transects, add = TRUE)}

```

To extract depth data from each of the transects I use QGIS in R below
0. Add the Hawaii Raster layer and Trimmed Transect layer.
1. Use 'Densify by interval' on the trimmed transects such that vertices are placed every 5 m
2. Use 'Extract vertices' from the densified layer to create a new point layer containing the vertices
3. Use the 'Sample raster values' to extract the depths at each vertex and then save this to a new shapefile.

```{r Extract Depths}
# install.package("remotes")
#remotes::install_github("JanCaha/r_package_qgis")
library(qgis)

#Sample along transects at 5 m intervals
result<-qgis::qgis_densifygeometriesgivenaninterval(INPUT = Hawaii_Transects,
                                                    INTERVAL = 5) %>% st_as_sf()

st_write(result,here("result.gpkg"), delete_layer=TRUE)

#Extract the vertices from the 5 m sampling
vertices<-qgis::qgis_extractvertices(result)%>% st_as_sf()

st_write(vertices,here("vertices.gpkg"), delete_layer=TRUE)

#Sample the raster at each of the 5 m points
sampled<-qgis::qgis_rastersampling(INPUT=vertices, RASTERCOPY = Hawaii_Bathymetry)%>% st_as_sf()

st_write(sampled,here("sampled.gpkg"), delete_layer=TRUE)
# Check if the directory exists
if (!dir.exists(here("2_output", "modified_rasters", "bathymetry"))) {
  # If it doesn't exist, create it
  dir.create(here("2_output", "modified_rasters", "bathymetry"), recursive = TRUE)
} 

#write the sampled data
st_write(sampled,here("2_output", "modified_shapefiles", "bathymetry", "sampled_transect_depths.gpkg"), delete_layer=TRUE)
```

```{r Extract two Transects}
library(sf)
library(raster)
library(here)
library(tidyverse)

sampled<- st_read(here("2_output", "modified_shapefiles", "bathymetry", "sampled_transect_depths.gpkg"))
sampled<- st_drop_geometry(sampled)

Characteristic_Reef <-sampled%>%
  filter(Transects_ID==284)%>%
  select(Transects_ID, distance, SAMPLE_1)%>%
  mutate(Distance=round(distance, digits=0),
         Depth=round(SAMPLE_1, digits=3),
         .keep="unused")%>%
  filter(is.na(Depth)!=TRUE)

Characteristic_Reef<-Characteristic_Reef%>%
  arrange(-Distance)

write_csv(Characteristic_Reef, "Characteristic_Reef.csv")

### Write a DEP file for XBeach

# Load your depth data
# This is an example, replace this with your actual data loading method
depth_data <- Characteristic_Reef$Depth  # a 10x10 matrix of random depth values

# Define the file path for the .dep file
file_path <- "Characteristic_Reef.dep"

# Open a connection to the file
file_conn <- file(file_path, "w")

# Write the depth data
write(paste(depth_data, collapse=" "), file_conn)

# Close the file connection
close(file_conn)


Characteristic_Reef %>%
  ggplot(aes(x=Distance, y=Depth))+
  geom_point()
  
Characteristic_Non_Reef <-sampled%>%
  filter(Transects_ID==1442)%>%
  select(Transects_ID, distance, SAMPLE_1)%>%
  mutate(Distance=round(distance, digits=0),
         Depth=round(SAMPLE_1, digits=3),
         .keep="unused")%>%
  filter(is.na(Depth)!=TRUE)

Characteristic_Non_Reef<-Characteristic_Non_Reef%>%
  arrange(-Distance)

write_csv(Characteristic_Non_Reef, "Characteristic_Non_Reef.csv")

Characteristic_Non_Reef %>%
  ggplot(aes(x=Distance, y=Depth))+
  geom_point()

```
